requirements.txt
django==4.2.15
djangorestframework==3.15.2
pytest-django==4.8.0
Faker==33.1.0